backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.0
pydantic>=2.0.0
aiofiles>=0.8.0
python-dotenv>=1.0.0
httpx>=0.24.0
pymongo>=4.4.0
